more strict static memory allocation limit

+++ doors.h
@@ -6,8 +6,8 @@
 #endif
 
 #include <limits.h>
-#if (NUMBER_OF_DOORS > UINT_MAX)
-#error (n) doors exceeds storage allocation limit for (unsigned int).
+#if (NUMBER_OF_DOORS > SHRT_MAX)
+#error (n) doors exceeds storage allocation limit for (short).
 #endif
 
 typedef struct {
